Split out Cassandra support into its own library
