+++ dev/archetypes/maven-java/src/main/resources/archetype-resources/__service1Name__-impl/pom.xml
@@ -22,7 +22,7 @@
         </dependency>
         <dependency>
             <groupId>com.lightbend.lagom</groupId>
-            <artifactId>lagom-javadsl-persistence_2.11</artifactId>
+            <artifactId>lagom-javadsl-persistence-cassandra_2.11</artifactId>
         </dependency>
         <dependency>
             <groupId>com.lightbend.lagom</groupId>
