+++ dev/archetypes/maven-java/src/main/resources/archetype-resources/pom.xml
@@ -60,7 +60,7 @@
             </dependency>
             <dependency>
                 <groupId>com.lightbend.lagom</groupId>
-                <artifactId>lagom-javadsl-persistence_2.11</artifactId>
+                <artifactId>lagom-javadsl-persistence-cassandra_2.11</artifactId>
                 <version>${lagom.version}</version>
             </dependency>
             <dependency>
