+++ dev/maven-plugin/src/maven-test/start-all/pom.xml
@@ -61,7 +61,7 @@
             </dependency>
             <dependency>
                 <groupId>com.lightbend.lagom</groupId>
-                <artifactId>lagom-javadsl-persistence_${scala.binary.version}</artifactId>
+                <artifactId>lagom-javadsl-persistence-cassandra_${scala.binary.version}</artifactId>
                 <version>${lagom.version}</version>
             </dependency>
             <dependency>
