+++ dev/sbt-plugin/src/sbt-test/sbt-plugin/distribution/build.sbt
@@ -9,7 +9,7 @@ lazy val lagomProj = (project in file("lagomProj")).enablePlugins(LagomJava)
   .settings(
     name := "lagom-dist-proj",
     version := "1.0-SNAPSHOT",
-    libraryDependencies ++= Seq(lagomJavadslPersistence, lagomSbtScriptedLibrary)
+    libraryDependencies ++= Seq(lagomJavadslPersistenceCassandra, lagomSbtScriptedLibrary)
   )
 
 lazy val playProj = (project in file("playProj")).enablePlugins(PlayScala, LagomPlay)
