+++ dev/sbt-plugin/src/sbt-test/sbt-plugin/run-services/build.sbt
@@ -2,4 +2,4 @@ lazy val root = (project in file(".")).enablePlugins(LagomJava)
 
 lagomUnmanagedServices in ThisBuild := Map("externalservice" -> "http://localhost:6000")
 
-libraryDependencies += lagomJavadslPersistence
+libraryDependencies += lagomJavadslPersistenceCassandra
