+++ persistence-cassandra/src/main/scala/com/lightbend/lagom/internal/persistence/cassandra/CassandraPersistentEntityRegistry.scala
