+++ persistence-cassandra/src/main/scala/com/lightbend/lagom/javadsl/persistence/cassandra/CassandraConfig.scala
@@ -1,0 +1,13 @@
+/*
+ * Copyright (C) 2016 Lightbend Inc. <http://www.lightbend.com>
+ */
+package com.lightbend.lagom.javadsl.persistence.cassandra
+
+import org.pcollections.PSet
+
+trait CassandraConfig {
+
+  /** Returns the Cassandra contact-points. */
+  def uris: PSet[CassandraContactPoint]
+
+}
