+++ persistence-cassandra/src/test/resources/reference.conf
