+++ persistence/src/main/scala/com/lightbend/lagom/javadsl/persistence/AggregateEventTag.scala
@@ -103,7 +103,7 @@ object AggregateEventTag {
    * @param entityId The ID of the entity.
    * @return The selected shard number.
    */
-  def selectShard(numShards: Int, entityId: String) = {
+  def selectShard(numShards: Int, entityId: String): Int = {
     Math.abs(entityId.hashCode) % numShards
   }
 
@@ -114,7 +114,7 @@ object AggregateEventTag {
    * @param shardNo The shard number.
    * @return The name of the shard tag.
    */
-  def shardTag(baseTagName: String, shardNo: Int) = {
+  def shardTag(baseTagName: String, shardNo: Int): String = {
     s"$baseTagName$shardNo"
   }
 }
