+++ persistence/src/main/scala/com/lightbend/lagom/javadsl/persistence/PersistenceModule.scala
@@ -3,31 +3,21 @@
  */
 package com.lightbend.lagom.javadsl.persistence
 
-import com.google.inject.{ AbstractModule, Inject, Key, TypeLiteral }
+import com.google.inject.{ AbstractModule, Inject, TypeLiteral }
 import com.google.inject.matcher.AbstractMatcher
 import com.google.inject.spi.InjectionListener
 import com.google.inject.spi.TypeEncounter
 import com.google.inject.spi.TypeListener
-import com.lightbend.lagom.internal.persistence.cassandra.CassandraPersistentEntityRegistry
-import com.lightbend.lagom.javadsl.persistence.cassandra.{ CassandraConfig, CassandraReadSide, CassandraSession }
-import com.lightbend.lagom.internal.persistence.cassandra.CassandraConfigProvider
-import com.lightbend.lagom.internal.persistence.cassandra.CassandraReadSideImpl
-import com.lightbend.lagom.internal.persistence.cassandra.CassandraSessionImpl
-import com.lightbend.lagom.internal.persistence.cassandra.ServiceLocatorHolder
 import com.lightbend.lagom.javadsl.api.ServiceLocator
 import com.lightbend.lagom.javadsl.persistence.PersistenceModule.InitServiceLocatorHolder
 import akka.actor.ActorSystem
-import com.lightbend.lagom.internal.persistence.ReadSideImpl
+import com.lightbend.lagom.internal.persistence.{ ReadSideImpl, ServiceLocatorHolder }
 
 /**
  * Guice module for the Persistence API.
  */
 class PersistenceModule extends AbstractModule {
   override def configure(): Unit = {
-    binder.bind(classOf[PersistentEntityRegistry]).to(classOf[CassandraPersistentEntityRegistry])
-    binder.bind(classOf[CassandraSession]).to(classOf[CassandraSessionImpl])
-    binder.bind(classOf[CassandraReadSide]).to(classOf[CassandraReadSideImpl])
-    binder.bind(classOf[CassandraConfig]).toProvider(classOf[CassandraConfigProvider])
     binder.bind(classOf[PersistenceModule.InitServiceLocatorHolder]).asEagerSingleton()
     binder.bind(classOf[ReadSide]).to(classOf[ReadSideImpl])
     initServiceLocatorHolder()
